requirements.txt
mysqlclient==2.1.1 ; python_version >= "3.11" and python_version < "4.0"
